REC6-  Organizando el código en carpetas
Se crea carpetas para models y routes.  Se organiza los archivos en
estas carpetas.
