+++ models/event.js
@@ -18,12 +18,17 @@ mongoose.connection.on('disconnected', () => {
 });
 
 const Schema = mongoose.Schema;
-const eventSchema = new Schema({
-  _id: { type: Number, required: true, unique: true },
-  title: { type: String, required: true },
-  description: { type: String, required: true },
-  date: { type: Date }
-});
+const eventSchema = new Schema(
+  {
+    _id: { type: Number, required: true, unique: true },
+    title: { type: String, required: true },
+    description: { type: String, required: true },
+    date: { type: Date }
+  },
+  {
+    versionKey: false
+  }
+);
 Event = mongoose.model('Event', eventSchema);
 
 module.exports = Event;
